Bug 519595 - NJ merge: avoid compile warning in NativeX64.cpp.  r=dvander.
UltraBlame original commit: 67417b2e65f74e38b1337ccea8c5e956a8218844

+++ js/src/nanojit/NativeX64.cpp
@@ -1154,7 +1154,7 @@ namespace nanojit
         if (kind == 0) {
             
             
-            if (a < NumArgRegs) {
+            if (a < (uint32_t)NumArgRegs) {
                 
                 prepResultReg(ins, rmask(argRegs[a]));
             } else {
